Создали компоненты «Список отзывов» и «Отзыв»
