+++ project/src/types/review.ts
@@ -1,0 +1,9 @@
+export type Review = {
+  userAvatar: string;
+  userName: string;
+  rating: string;
+  text: string;
+  dateAdd: string;
+};
+
+export type Reviews = Review[];
